+++ packages/api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @yukinu/api
 
+## 1.0.2
+
+### Patch Changes
+
+- 58a6475: bump
+
+## 1.0.1
+
+### Patch Changes
+
+- 5ae2461: fix calculate payment amount
+
 ## 1.0.0
 
 ### Major Changes